requirements.test.txt
# Strictly for tests
pytest<6.0
pytest-cov<3.0.0
pytest-homeassistant-custom-component
# From our manifest.json for our custom component
aiohttp_cors==0.7.0
paho-mqtt==1.5.1
